apps/dfl/compose-gen/requirements.txt
ruamel.yaml~=0.18
gpustat~=1.1